requirements.txt
pytesseract==0.3.10
PyMuPDF==1.24.0
google-cloud-vision==3.7.2
numpy